requirements.txt
audioop-lts==0.2.2
Flask==3.1.2
gTTS==2.5.4
pyaudio==0.2.14
pydub==0.25.1
requests==2.32.5
python-dotenv==1.1.1
numpy==2.3.3